requirements.txt
ultralytics
#gradio
pillow
fastapi
uvicorn
torch
requests
python-multipart

